Update patch set 1
Patch Set 1:

> What is PAD? 
Please see prior replies on this change list.

> This commit will not break old models.. we detect if the new model api is being used via /model/methods/* in the url path. If there is some common library that would be good to use npm possibly as well so people can create web demos without copy and paste. 

To be technically accurate, it's more complex than just providing a URL.  With the new runners, the proto should be retrieved from the runner to make sure that it's accurate and matches the endpoints.  re NPM, making another webpage client library may be useful but is way out of scope.  The code is already duplicated in the three repos, but the modernized update was to consolidate in PAD.

Patch-set: 1