Update patch set 1
Patch Set 1:

It won't break the old _models_, but it does make whatever code communicates with models (such as this) incompatible.

I think Justin's suggestion is to write this communication code to support both models with old-style model runners and new-style ones for backward compatibility.

I for one would vote for just biting the bullet and obsoleting code that interfaces with old-style model runners.

Patch-set: 1